move view traits into a separate header
